+++ src/main/java/com/nbcamp/mypocketbookapi/dto/content/ContentCreateRequestDto.java
@@ -1,13 +1,10 @@
-package com.nbcamp.mypocketbookapi.dto;
+package com.nbcamp.mypocketbookapi.dto.content;
 
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
-public class ContentSearchRequestDto {
-    // long id는 아직 없기때문에 생략
-    private Long memberId;
+public class ContentCreateRequestDto {
+
     private String isbn;
     private String title;
     private String thumbnail;
