+++ src/main/java/com/nbcamp/mypocketbookapi/dto/content/ContentSearchRequestDto.java
@@ -1,12 +1,11 @@
-package com.nbcamp.mypocketbookapi.dto;
+package com.nbcamp.mypocketbookapi.dto.content;
 
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
-public class ContentCreateRequestDto {
-
+public class ContentSearchRequestDto {
+    // long id는 아직 없기때문에 생략
+    private Long memberId;
     private String isbn;
     private String title;
     private String thumbnail;
