+++ src/main/java/com/nbcamp/mypocketbookapi/repository/ContentJpaRepository.java
@@ -5,10 +5,13 @@ import com.nbcamp.mypocketbookapi.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface ContentJpaRepository extends JpaRepository<Content, Long> {
 
     List<Content> findByMember(Member member);
 
     boolean existsByIsbn(String isbn);
+
+    Optional<Content> findByIdAndMemberId(Long contentId, Long memberId);
 }
